game editor

+++ Source/Engine/GfxDriver/Buffer.hpp
@@ -9,7 +9,7 @@ public:
     struct CreateInfo
     {
         BufferUsageFlags usages;
-        uint32_t size;
+        size_t size;
         bool visibleInCPU = false;
         const char* debugName = nullptr;
     };
@@ -18,6 +18,6 @@ public:
     virtual ~Buffer(){};
     virtual void* GetCPUVisibleAddress() = 0;
     virtual void SetDebugName(const char* name) = 0;
-    virtual uint32_t GetSize() = 0;
+    virtual size_t GetSize() = 0;
 };
 } // namespace Engine::Gfx
